% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MVA_module.R
\name{MVA_module}
\alias{MVA_module}
\title{SpectroPipeR: MVA module}
\usage{
MVA_module(
  SpectroPipeR_data_quant = NULL,
  HCPC_analysis = FALSE,
  costum_colors = NULL
)
}
\arguments{
\item{SpectroPipeR_data_quant}{it is the SpectroPipeR_data_quant list object from norm_quant_module() object e.g. \code{data_input_norm_quant} see example below}

\item{HCPC_analysis}{boolean; should a HCPC be performed or not}

\item{costum_colors}{if you would like to use your own colors please provide a named color vector (e.g. c(condition1 = "black", condition2 = "grey")); names should have the same naming and length like the conditions set in Spectronaut}
}
\value{
saves MVA analysis results in output folder

SpectroPipeR_data_MVA list object with the MVA analysis data in addition to the automatically saved tables and plots
For the description of the generated figures and tables please read the manual & vignettes\tabular{ll}{
   \if{html}{\out{<u>}} \strong{list element} \if{html}{\out{</u>}} \tab \if{html}{\out{<u>}} \strong{description} \if{html}{\out{</u>}} \cr
   PCA_peptide_intensity \tab \emph{PCA list element:} PCA list element of peptide int. \cr
   PCA_protein_intensity \tab \emph{PCA list element:} PCA list element of protein int. \cr
   UMAP_protein_intensity \tab \emph{umap element:} UMAP element of protein int. \cr
   peptide_intensity_correlation \tab \emph{matrix:} Spearman correlation scores of peptide int. \cr
   protein_intensity_correlation \tab \emph{matrix:} Spearman correlation scores of protein int. \cr
}
}
\description{
Function for conducting multivariate analysis, which represents the third step in the pipeline and builds upon the normalization and quantification module from step 2.
}
\examples{
#load library
library(SpectroPipeR)

# parameter list
params <- list(output_folder = "../SpectroPipeR_test_folder",
              ion_q_value_cutoff = 0.01,
              id_drop_cutoff = 0.3,
              normalization_method = "median",
              normalization_factor_cutoff_outlier = 4,
              filter_oxidized_peptides = TRUE,
              protein_intensity_estimation = "MaxLFQ",
              stat_test = "modt",
              type_slr = "median",
              fold_change = 1.5,
              p_value_cutoff = 0.05,
              paired = FALSE
)

# example input file
example_file_path <- system.file("extdata",
                                "SN_test_HYE_mix_file.tsv",
                                package="SpectroPipeR")

# step 1: load Spectronaut data module
SpectroPipeR_data <- read_spectronaut_module(file = example_file_path,
                                            parameter = params,
                                            print.plot = FALSE)

# step 2: normalize & quantification module
SpectroPipeR_data_quant <- norm_quant_module(SpectroPipeR_data = SpectroPipeR_data)

# step 3: MVA module
SpectroPipeR_MVA <- MVA_module(SpectroPipeR_data_quant = SpectroPipeR_data_quant,
          HCPC_analysis = FALSE)
}
