% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpectroPipeR_logo_export.R
\name{SpectroPipeR_logo_export}
\alias{SpectroPipeR_logo_export}
\title{SpectroPipeR - logo export}
\usage{
SpectroPipeR_logo_export(output_location = "")
}
\arguments{
\item{output_location}{folder location where the SpectroPipeR logo should be copied}
}
\value{
copies the SpectroPipeR logo to the designated folder
}
\description{
SpectroPipeR - logo export
}
\examples{

SpectroPipeR_logo_export(output_location = "../SpectroPipeR_test_folder")

}
