% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpectroPipeR_gui.R
\name{SpectroPipeR_ui}
\alias{SpectroPipeR_ui}
\title{SpectroPipeR UI function}
\usage{
SpectroPipeR_ui()
}
\value{
launches a user interface in the browser
}
\description{
launches a user interface in the browser, providing access to the majority of SpectroPipeR's functionalities.
}
\examples{
\donttest{
SpectroPipeR_ui()
}

}
