% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics_module.R
\name{statistics_module}
\alias{statistics_module}
\title{SpectroPipeR: statistics module}
\usage{
statistics_module(
  SpectroPipeR_data_quant = NULL,
  condition_comparisons = NULL,
  number_of_cores = 2
)
}
\arguments{
\item{SpectroPipeR_data_quant}{it is the SpectroPipeR_data_quant list object from norm_quant_module() object e.g. \code{SpectroPipeR_data_quant} see example below}

\item{condition_comparisons}{condition comparisons for pairwise- comparison; e.g. condition_comparisons <- cbind(c("condition1","control"),c("condition3","control") )}

\item{number_of_cores}{number of processor cores to be used for the calculations default = 2;
\code{parallel::detectCores()-2} for faster processing (will detect the number of cores in the system and use nearly all cores)}
}
\value{
SpectroPipeR_statistics list element containing the statistics analysis results in addition to the automatically generated plots and tables in output folder
For the description of the generated figures and tables please read the manual & vignettes\tabular{ll}{
   \if{html}{\out{<u>}} \strong{list element} \if{html}{\out{</u>}} \tab \if{html}{\out{<u>}} \strong{description} \if{html}{\out{</u>}} \cr
   stat_results \tab \emph{tibble:} statistical analysis results table \cr
   stat_column_description \tab \emph{tibble:} statistical analysis results table column description \cr
   stats_results_iBAQ_quantiles \tab \emph{tibble:} statistical analysis results table containing \cr
    \tab the iBAQ quantilies (Q1-Q10) of the protein per group for a better \cr
    \tab ratio judgement \cr
   stat_results_filtered \tab \emph{tibble:} filtered (user defined FC and adj. p-value) statistical \cr
    \tab analysis results table \cr
}
}
\description{
Function for performing statistical analysis on Spectronaut output reports, representing the fourth step in the pipeline and is dependent on the normalization and quantification module output from step 2.
}
\examples{
\donttest{
#load library
library(SpectroPipeR)

# use default parameters list
params <- list(output_folder = "../SpectroPipeR_test_folder")

# example input file
example_file_path <- system.file("extdata",
                                "SN_test_HYE_mix_file.tsv",
                                package="SpectroPipeR")

# step 1: load Spectronaut data module
SpectroPipeR_data <- read_spectronaut_module(file = example_file_path,
                                            parameter = params,
                                            print.plot = FALSE)

# step 2: normalize & quantification module
SpectroPipeR_data_quant <- norm_quant_module(SpectroPipeR_data = SpectroPipeR_data)

# step 3: MVA module
SpectroPipeR_data_MVA <- MVA_module(SpectroPipeR_data_quant = SpectroPipeR_data_quant,
          HCPC_analysis = FALSE)

# step 4: statistics module
SpectroPipeR_data_stats <- statistics_module(SpectroPipeR_data_quant = SpectroPipeR_data_quant,
                                            condition_comparisons = cbind(c("HYE mix A",
                                                                            "HYE mix B")
                                                                          )
                                            )
}
}
