% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XIC_plot_module.R
\name{XIC_plot_module}
\alias{XIC_plot_module}
\title{Extraction and plotting of XICs from Spectronaut's exported SQLite Databases (Spectronaut version >  18.7.240506.55695;  XIC DB export generates one SQLite db file per raw file)}
\usage{
XIC_plot_module(
  Spectronaut_report_path = NULL,
  Spectronaut_xicDB_path = NULL,
  protein_groups = NULL,
  output_path = NULL,
  export_csv_files = FALSE,
  run_specific_y_axis = FALSE,
  selected_conditions = NULL,
  number_of_cores = 2
)
}
\arguments{
\item{Spectronaut_report_path}{file; Path pointing to Spectronaut analysis report using SpectroPipeR report export scheme.}

\item{Spectronaut_xicDB_path}{folder; Path pointing to Spectronaut all XIC SQLite database export folder from that analysis (use pipeline mode in Spectronaut).}

\item{protein_groups}{Vector containing user-selected protein group IDs for generating plots. e.g. c("P38720","P29311")}

\item{output_path}{Define the output directory for the generated ion-specific plots (PDF) and data tables (CSV) for user-selected protein groups.}

\item{export_csv_files}{Export individual CSV files per ion (boolean, default: FALSE).}

\item{run_specific_y_axis}{if the XIC plot should be globally scaled y-axis or run specific (default: FALSE)}

\item{selected_conditions}{Vector of user-selected conditions for plotting (serves for both selection and ordering).}

\item{number_of_cores}{Define the number of CPU cores to utilize for parallel XIC extraction and plotting.}
}
\value{
Exports user-selected protein group XIC data (MS1/MS2) in both PDF and CSV formats for all detected ions of a protein.
}
\description{
Exports user-selected protein group XIC (MS1/MS2) data in both PDF and CSV formats for all detected ions.
}
\details{
Description of the Spectronaut scores shown in the output plots:\tabular{ll}{
   \if{html}{\out{<u>}} \strong{parameter} \if{html}{\out{</u>}} \tab \if{html}{\out{<u>}} \strong{description} \if{html}{\out{</u>}} \cr
   EG.Qvalue \tab The Q-value is the multiple test corrected p-value. The Q-value can be used \cr
    \tab as an estimate for the FDR at a given Cscore cutoff. \cr
   EG.PEP \tab The posterior error probability for a peptide precursor observation. The PEP \cr
    \tab is calculated as the decoy probability function for a specficic Cscore divided \cr
    \tab by the probability function for decoy plus target. PEP = d/(t+d) where d is the \cr
    \tab value of the decoy probability function for a given Cscore and t is the value for \cr
    \tab target probability function for a given Cscore. \cr
   FG.ShapeQualityScore \tab The average of all available peak quality measurements \cr
    \tab (scale: 1.0 \code{best} to -1.0 \code{worst}) \cr
   EG.IntCorrScore \tab The correlation score between the relative fragment intensities as provided \cr
    \tab by the spectral librarxy and the empirical observed fragtment intensitiers \cr
    \tab at XIC peak apex. \cr
   EG.SignalToNoise \tab The signal to noise ratio of the LC-MS peak for a given peptide precursor. \cr
    \tab The signal is calc. as the maximum intensity of the fragment sum XIC within \cr
    \tab the peak boundaries. The noise is calculated as the average fragment sum XIC \cr
    \tab intensity outside the peak boundaries. \cr
   EG.Cscore \tab The discriminant score used for identification. This score is a linear combination \cr
    \tab of all applicable scores for a certain workflow. \cr
   EG.Identified \tab Returns weather a precursor is considered identified or not. In order to be \cr
    \tab identified it has to match both the precursor and the protein FDR cutoff. \cr
   dppp (MS1) \tab Returns the number of data points (scans) that make up the MS1 XIC of the selected \cr
    \tab LC-MS peak for this precursor. Only data points that lie within the start and the \cr
    \tab end RT are considered. \cr
   dppp (MS2) \tab Returns the number of data points (scans) that make up the MS2 XIC of the selected \cr
    \tab LC-MS peak for this precursor. Only data points that lie within the start and the \cr
    \tab end RT are considered. \cr
}


The vertical dashed line in the plots indicate the individual ApexRT, which is calculated based on all fragment XIC for a specific peptide precursor ion.
}
\examples{
\donttest{
#setup example input paths / protein selection for plotting
Spectronaut_report_path <- system.file("extdata/HYE_demo_data",
                                       "HYE_demo_data_Report_SpectroPipeR.tsv",
                                        package="SpectroPipeR")
Spectronaut_xicDB_path <- system.file("extdata/HYE_demo_data/XIC_DBs",package="SpectroPipeR")
protein_groups <- c("P29311","P38720")
output_path <- "../SpectroPipeR_test_folder/single_XIC_plots"

# extracting and plotting of XIC
XIC_plot_module(Spectronaut_report_path = Spectronaut_report_path,
                Spectronaut_xicDB_path = Spectronaut_xicDB_path,
                protein_groups = protein_groups,
                output_path = output_path,
                number_of_cores = 2
)
}
}
